get type by request (advanced): including type has contract almost expiring, contract has status reversed

+++ src/main/java/org/avengers/capstone/hostelrenting/model/Contract.java
@@ -74,7 +74,7 @@ public class Contract {
      * contract duration
      */
     @Column(nullable = false)
-    private Integer duration;
+    private Float duration;
 
     /**
      * start time of duration
